InsuranceContracts.jl

LifeInsuranceDataModel.InsuranceContracts.Contract — Type

Contract

a contract component of a bitemporal entity

LifeInsuranceDataModel.InsuranceContracts.ContractPartnerRef — Type

ContractPartnerRef

a partner reference of a contract component, i.e. policy holder, premium payer

LifeInsuranceDataModel.InsuranceContracts.ContractPartnerRefRevision — Type

ContractPartnerRefRevision

a revision of a contract's partner reference

LifeInsuranceDataModel.InsuranceContracts.ContractPartnerRole — Type

ContractPartnerRole

role e.g. policy holder or premium payer

LifeInsuranceDataModel.InsuranceContracts.ContractRevision — Type

ContractRevision

a revision of a contract component

LifeInsuranceDataModel.InsuranceContracts.ProductItem — Type

ProductItem

a productitem component of a contract component

LifeInsuranceDataModel.InsuranceContracts.ProductItemRevision — Type

ProductItemRevision

a revision of a productitem component

LifeInsuranceDataModel.InsuranceContracts.TariffItem — Type

TariffItem

a reference to a tariff with contractual parameters

LifeInsuranceDataModel.InsuranceContracts.TariffItemPartnerRef — Type

TariffItemPartnerRef

a reference to a partner of a tariffitem, i.e. insured person

LifeInsuranceDataModel.InsuranceContracts.TariffItemPartnerRefRevision — Type

TariffItemPartnerRefRevision

a revision of a productItem's partner reference

LifeInsuranceDataModel.InsuranceContracts.TariffItemPartnerRole — Type

TariffItemPartnerRole

role of Partner for tariffitem, 1. or 2. insured person

LifeInsuranceDataModel.InsuranceContracts.TariffItemRevision — Type

TariffItemRevision

a revision of a tariffitem

LifeInsuranceDataModel.InsuranceContracts.TariffItemRole — Type

TariffItemRole

role e.g. main or supplemental risk like life and occupational disabilty

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::ContractPartnerRefRevision) :: Type{ContractPartnerRef}

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::ContractRevision) :: Type{Contract}

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::ProductItemRevision) :: Type{ProductItem}

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::TariffItemPartnerRefRevision) :: Type{TariffItemPartnerRef}

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::TariffItemRevision) :: Type{TariffItem}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::ContractPartnerRef) :: Type{ContractPartnerRefRevision}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::Contract) :: Type{ContractRevision}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::ProductItem) :: Type{ProductItemRevision}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::TariffItemPartnerRef) :: Type{TariffItemPartnerRefRevision}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::TariffItem) :: Type{TariffItemRevision}

BitemporalPostgres.revisionTypes — Method

BitemporalPostgres.revisionTypes(entity::Val{:Contract}) defining the ComponentRevision types occurring in Contracts