InsuranceProducts.jl

LifeInsuranceDataModel.InsuranceProducts.Product — Type

Product

a component of a bitemporal entity

LifeInsuranceDataModel.InsuranceProducts.ProductPart — Type

ProductPart

the relation between a product and it's component tariffs

LifeInsuranceDataModel.InsuranceProducts.ProductPartRevision — Type

ProcuctPart_Revision

a revision of a ProductPart component of a bitemporal entity

LifeInsuranceDataModel.InsuranceProducts.ProductPartRole — Type

ProductPartRole

role e.g. main or supplemental risk like life and occupational disabilty, profit participation

LifeInsuranceDataModel.InsuranceProducts.ProductRevision — Type

Procuct_Revision

a revision of a Product component of a bitemporal entity

BitemporalPostgres.revisionTypes — Method

BitemporalPostgres.revisionTypes(entity::Val{:Product}) defining the ComponentRevision types occurring in Contracts