InsuranceTariffs.jl

LifeInsuranceDataModel.InsuranceTariffs.Tariff — Type

Tariff

a component of a bitemporal entity

LifeInsuranceDataModel.InsuranceTariffs.TariffPartnerRole — Type

TariffPartnerRole

a subcomponent of a tariff entity enumerationg the roles of tariff partners

LifeInsuranceDataModel.InsuranceTariffs.TariffPartnerRoleRevision — Type

TariffPartnerRoleRevision

a partner role revision of a Tariff component of a bitemporal entity

LifeInsuranceDataModel.InsuranceTariffs.TariffRevision — Type

Tariff_Revision

a revision of a Tariff component of a bitemporal entity

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::TariffPartnerRoleRevision) :: Type{TariffPartnerRole}

BitemporalPostgres.get_typeof_component — Method

BitemporalPostgres.gettypeofcomponent(revision::TariffRevision) :: Type{Tariff}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::TariffPartnerRole) :: Type{TariffPartnerRoleRevision}

BitemporalPostgres.get_typeof_revision — Method

BitemporalPostgres.gettypeofrevision(component::Tariff) :: Type{TariffRevision}

BitemporalPostgres.revisionTypes — Method

BitemporalPostgres.revisionTypes(entity::Val{:Tariff}) defining the ComponentRevision types occurring in Contracts